Remove unnecessary exit handling
I spent all that time reading NMS to figure out how portal behavior changed last time and forgot all of it by the time I returned. Noting it for when I inevitably forget why it's unnecessary and that the old bug where you'd be "rescued" from a portal across the map is handled.