Clean up logging a little
This is a bit of a hack, but modifying the configuration to not log the logger name for modules similarly to how plugin logger names are not logged is kind of a hassle. The vast majority of users will not be using custom log4j configs (and don't know how) which is also why Bukkit plugins traditionally have their own debug flags rather than letting the logger deal with it.