Extract allowBuild/allowBreak to helper
These calls have been in a bit of limbo since the ClaimPermissionCheckEvent was added. It's nice to reduce duplicate code, but all events being either a BlockPlaceEvent or BlockBreakEvent reduced the usability of the event funnel. Additionally, since event constructors are not considered API, moving away from them is a good idea.